Moved menu data interface into base menu class since theres no good reason to keep it in its own file
